Extract backend selection/hourly services for adapter reuse
