+++ src/backend/services/__init__.py
@@ -1,5 +1,27 @@
+from src.backend.services.hourly_payload_service import build_hourly_payload_for_resort
+from src.backend.services.resort_selection_service import (
+    apply_catalog_filters,
+    available_filters,
+    build_empty_payload,
+    catalog_item_with_display_name,
+    default_applied_filters,
+    load_supported_resort_catalog,
+    select_resorts_from_query,
+    split_query_values,
+    supported_catalog,
+)
 from src.backend.services.weather_service import build_weather_payload
 
 __all__ = [
+    "apply_catalog_filters",
+    "available_filters",
+    "build_empty_payload",
+    "build_hourly_payload_for_resort",
     "build_weather_payload",
+    "catalog_item_with_display_name",
+    "default_applied_filters",
+    "load_supported_resort_catalog",
+    "select_resorts_from_query",
+    "split_query_values",
+    "supported_catalog",
 ]
