+++ src/backend/weather_data_server.py
@@ -13,18 +13,20 @@ from urllib.parse import parse_qs, urlparse
 if str(Path(__file__).resolve().parents[2]) not in sys.path:
     sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-from src.backend.cache import JsonCache, ResortCoordinateCache, dated_cache_path
-from src.backend.airport_catalog import find_nearby_airports, load_airport_catalog
-from src.backend.constants import COORDINATES_CACHE_FILE
-from src.backend.io import seed_coordinate_cache_from_entries
-from src.backend.compute.payload_metadata import build_payload_metadata
-from src.backend.open_meteo import fetch_hourly_forecast, geocode
 from src.backend.pipelines.live_pipeline import run_live_payload
-from src.backend.resort_catalog import load_resort_catalog, search_resort_catalog
+from src.backend.services.hourly_payload_service import build_hourly_payload_for_resort
+from src.backend.services.resort_selection_service import (
+    apply_catalog_filters,
+    available_filters,
+    build_empty_payload,
+    catalog_item_with_display_name,
+    default_applied_filters,
+    load_supported_resort_catalog,
+    select_resorts_from_query,
+    split_query_values,
+    supported_catalog,
+)
 from src.shared.config import DEFAULT_RESORTS_FILE
-
-_SUPPORTED_PASS_TYPES = {"epic", "ikon"}
-
 
 def _append_common_headers(handler: BaseHTTPRequestHandler, allow_origin: str) -> None:
     handler.send_header("Access-Control-Allow-Origin", allow_origin)
@@ -33,64 +35,19 @@ def _append_common_headers(handler: BaseHTTPRequestHandler, allow_origin: str) -
 
 
 def _split_query_values(values: List[str], *, to_upper: bool = False) -> List[str]:
-    out: List[str] = []
-    for raw in values:
-        for part in raw.split(","):
-            val = part.strip().upper() if to_upper else part.strip().lower()
-            if val:
-                out.append(val)
-    seen = set()
-    return [x for x in out if not (x in seen or seen.add(x))]
-
-
-def _to_bool_flag(raw: str) -> bool:
-    return (raw or "").strip().lower() in {"1", "true", "yes", "on"}
+    return split_query_values(values, to_upper=to_upper)
 
 
 def _supported_catalog(catalog: List[Dict[str, object]]) -> List[Dict[str, object]]:
-    out: List[Dict[str, object]] = []
-    for item in catalog:
-        raw_pass_types = item.get("pass_types")
-        if not isinstance(raw_pass_types, list):
-            continue
-        pass_types = {str(v).strip().lower() for v in raw_pass_types if str(v).strip()}
-        if pass_types.intersection(_SUPPORTED_PASS_TYPES):
-            out.append(item)
-    return out
+    return supported_catalog(catalog)
 
 
 def _catalog_item_with_display_name(item: Dict[str, object]) -> Dict[str, object]:
-    out = dict(item)
-    out["display_name"] = str(item.get("display_name") or item.get("query") or "").strip()
-    out["website"] = str(item.get("website") or "").strip()
-    return out
+    return catalog_item_with_display_name(item)
 
 
 def _available_filters(catalog: List[Dict[str, object]]) -> Dict[str, Dict[str, int]]:
-    pass_type_counts: Dict[str, int] = {}
-    region_counts: Dict[str, int] = {}
-    subregion_counts: Dict[str, int] = {}
-    country_counts: Dict[str, int] = {}
-    for item in catalog:
-        region = str(item.get("region", "")).strip().lower()
-        if region:
-            region_counts[region] = region_counts.get(region, 0) + 1
-        subregion = str(item.get("subregion", "")).strip().lower()
-        if subregion:
-            subregion_counts[subregion] = subregion_counts.get(subregion, 0) + 1
-        country = str(item.get("country", "")).strip().upper()
-        if country:
-            country_counts[country] = country_counts.get(country, 0) + 1
-        for pass_type in item.get("pass_types", []) if isinstance(item.get("pass_types"), list) else []:
-            pt = str(pass_type).strip().lower()
-            if pt:
-                pass_type_counts[pt] = pass_type_counts.get(pt, 0) + 1
-    return {
-        "pass_type": pass_type_counts,
-        "region": region_counts,
-        "subregion": subregion_counts,
-        "country": country_counts,
-    }
+    return available_filters(catalog)
 
 
 def _apply_catalog_filters(
@@ -102,136 +59,22 @@ def _apply_catalog_filters(
     countries: List[str],
     search: str,
 ) -> List[Dict[str, object]]:
-    items = search_resort_catalog(catalog, search)
-    if pass_types:
-        allowed = set(pass_types)
-        items = [
-            item
-            for item in items
-            if allowed.intersection(
-                {str(v).strip().lower() for v in (item.get("pass_types") or []) if str(v).strip()}
-            )
-        ]
-    if region:
-        want = region.lower()
-        items = [item for item in items if str(item.get("region", "")).strip().lower() == want]
-    if subregions:
-        allowed_subregions = set(subregions)
-        items = [item for item in items if str(item.get("subregion", "")).strip().lower() in allowed_subregions]
-    if countries:
-        allowed_countries = set(countries)
-        items = [item for item in items if str(item.get("country", "")).strip().upper() in allowed_countries]
-    return items
+    return apply_catalog_filters(
+        catalog,
+        pass_types=pass_types,
+        region=region,
+        subregions=subregions,
+        countries=countries,
+        search=search,
+    )
 
 
 def _empty_payload(cache_file: str, geocode_cache_hours: int, forecast_cache_hours: int) -> Dict[str, object]:
-    return build_payload_metadata(
-        cache_path=cache_file,
-        cache_hits=0,
-        cache_misses=0,
-        geocode_cache_hours=geocode_cache_hours,
-        forecast_cache_hours=forecast_cache_hours,
-        reports=[],
-        failed=[],
-    )
+    return build_empty_payload(cache_file, geocode_cache_hours, forecast_cache_hours)
 
 
 def _default_applied_filters() -> Dict[str, object]:
-    return {
-        "pass_type": [],
-        "region": "",
-        "subregion": [],
-        "country": [],
-        "search": "",
-        "search_all": True,
-        "include_default": True,
-        "include_all": False,
-    }
-
-
-def select_resorts_from_query(qs: dict) -> tuple[List[str], str, dict, dict, bool]:
-    resorts = [x.strip() for x in qs.get("resort", []) if x.strip()]
-    pass_types = _split_query_values(qs.get("pass_type", []))
-    region = (qs.get("region", [""])[0] or "").strip().lower()
-    subregions = _split_query_values(qs.get("subregion", []))
-    countries = _split_query_values(qs.get("country", []), to_upper=True)
-    search_text = (qs.get("search", [""])[0] or "").strip()
-    has_search_all = "search_all" in qs
-    search_all = _to_bool_flag((qs.get("search_all", [""])[0] or "")) if has_search_all else True
-    has_include_default = "include_default" in qs
-    include_default = _to_bool_flag((qs.get("include_default", [""])[0] or "")) if has_include_default else False
-    include_all = _to_bool_flag((qs.get("include_all", [""])[0] or ""))
-    applied = {
-        "pass_type": pass_types,
-        "region": region,
-        "subregion": subregions,
-        "country": countries,
-        "search": search_text,
-        "search_all": search_all,
-        "include_default": include_default,
-        "include_all": include_all,
-    }
-
-    catalog = _supported_catalog(load_resort_catalog(DEFAULT_RESORTS_FILE))
-    available = _available_filters(catalog)
-    has_filters = bool(
-        pass_types or region or subregions or countries or search_text or include_all or has_include_default or has_search_all
-    )
-    if not has_filters:
-        applied["search_all"] = True
-        applied["include_default"] = not bool(resorts)
-        applied["include_all"] = False
-        resorts_file = "" if resorts else DEFAULT_RESORTS_FILE
-        return resorts, resorts_file, applied, available, False
-
-    if search_text and search_all:
-        # Search-all mode ignores pass/region/country/default scope and searches full supported catalog.
-        filtered_catalog = search_resort_catalog(catalog, search_text)
-    elif include_default:
-        default_catalog = [item for item in catalog if bool(item.get("default_enabled", False))]
-        filtered_catalog = _apply_catalog_filters(
-            default_catalog,
-            pass_types=pass_types,
-            region=region,
-            subregions=subregions,
-            countries=countries,
-            search=search_text,
-        )
-    elif include_all and not (pass_types or region or subregions or countries or search_text):
-        filtered_catalog = list(catalog)
-    else:
-        filtered_catalog = _apply_catalog_filters(
-            catalog,
-            pass_types=pass_types,
-            region=region,
-            subregions=subregions,
-            countries=countries,
-            search=search_text,
-        )
-
-    allowed_queries = {
-        str(item.get("query", "")).strip()
-        for item in filtered_catalog
-        if str(item.get("query", "")).strip()
-    }
-    if resorts:
-        selected = [r for r in resorts if r in allowed_queries]
-    else:
-        selected = [
-            str(item["query"]).strip()
-            for item in filtered_catalog
-            if str(item.get("query", "")).strip()
-        ]
-    no_match = len(selected) == 0
-    return selected, "", applied, available, no_match
-
-
-def _parse_hours(raw: str, default: int = 72) -> int:
-    try:
-        value = int(raw)
-    except (TypeError, ValueError):
-        value = default
-    return max(1, min(240, value))
+    return default_applied_filters()
 
 
 def _hourly_payload_for_resort(
@@ -242,88 +85,21 @@ def _hourly_payload_for_resort(
     geocode_cache_hours: int,
     forecast_cache_hours: int,
 ) -> Dict[str, object] | None:
-    catalog = _supported_catalog(load_resort_catalog(DEFAULT_RESORTS_FILE))
-    item = next((r for r in catalog if str(r.get("resort_id", "")) == resort_id), None)
-    if item is None:
-        return None
-
-    cache_path = dated_cache_path(cache_file)
-    cache = JsonCache(cache_path)
-    coord_cache = ResortCoordinateCache(COORDINATES_CACHE_FILE)
-    query = str(item.get("query", "")).strip()
-    seed_coordinate_cache_from_entries(coord_cache, [item])
-    location = geocode(
-        query,
-        cache=cache,
-        ttl_seconds=geocode_cache_hours * 3600,
-        coord_cache=coord_cache,
-    )
-    if location is None:
-        return {
-            "error": f"Unable to geocode resort '{query}'",
-            "resort_id": resort_id,
-            "query": query,
-        }
-
-    forecast = fetch_hourly_forecast(
-        location,
-        cache=cache,
-        ttl_seconds=forecast_cache_hours * 3600,
+    return build_hourly_payload_for_resort(
+        resort_id=resort_id,
         hours=hours,
+        cache_file=cache_file,
+        geocode_cache_hours=geocode_cache_hours,
+        forecast_cache_hours=forecast_cache_hours,
     )
-    cache.save()
-    coord_cache.save()
+
+
+def _parse_hours(raw: str, default: int = 72) -> int:
     try:
-        airports = load_airport_catalog()
-    except Exception:
-        airports = []
-    nearby_airports = find_nearby_airports(
-        resort_latitude=location.latitude,
-        resort_longitude=location.longitude,
-        airports=airports,
-        radius_miles=250.0,
-    )
-
-    hourly = forecast.get("hourly", {}) if isinstance(forecast, dict) else {}
-    times = list(hourly.get("time", []))
-    n = min(hours, len(times))
-    metric_keys = [
-        "snowfall",
-        "rain",
-        "precipitation_probability",
-        "snow_depth",
-        "wind_speed_10m",
-        "wind_direction_10m",
-        "visibility",
-    ]
-    trimmed_hourly: Dict[str, object] = {"time": times[:n]}
-    for key in metric_keys:
-        values = hourly.get(key, [])
-        if isinstance(values, list):
-            trimmed_hourly[key] = values[:n]
-        else:
-            trimmed_hourly[key] = []
-
-    return {
-        "resort_id": resort_id,
-        "query": query,
-        "display_name": str(item.get("display_name") or query).strip(),
-        "website": str(item.get("website") or "").strip(),
-        "matched_name": location.name,
-        "country": item.get("country"),
-                "region": item.get("region"),
-                "subregion": item.get("subregion"),
-                "pass_types": item.get("pass_types", []),
-        "timezone": forecast.get("timezone"),
-        "model": "ecmwf_ifs025",
-        "input_latitude": location.latitude,
-        "input_longitude": location.longitude,
-        "resolved_latitude": forecast.get("latitude"),
-        "resolved_longitude": forecast.get("longitude"),
-        "nearby_airports": nearby_airports,
-        "hours": n,
-        "hourly": trimmed_hourly,
-    }
+        value = int(raw)
+    except (TypeError, ValueError):
+        value = default
+    return max(1, min(240, value))
 
 
 def make_handler(
@@ -372,7 +148,7 @@ def make_handler(
                     return
                 hours = _parse_hours((qs.get("hours", ["72"])[0] or "72"), default=72)
                 try:
-                    result = _hourly_payload_for_resort(
+                    result = build_hourly_payload_for_resort(
                         resort_id=resort_id,
                         hours=hours,
                         cache_file=cache_file,
@@ -393,7 +169,7 @@ def make_handler(
             if parsed.path == "/api/resorts":
                 search_text = (qs.get("search", [""])[0] or "").strip()
                 try:
-                    catalog = _supported_catalog(load_resort_catalog(DEFAULT_RESORTS_FILE))
+                    catalog = load_supported_resort_catalog(DEFAULT_RESORTS_FILE)
                 except Exception as exc:
                     self._write_json(500, {"error": f"Failed to load resort catalog: {exc}"})
                     return
