+++ tests/backend/test_weather_data_server_filters.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from src.backend.weather_data_server import DEFAULT_RESORTS_FILE, select_resorts_from_query
+from src.backend.services.resort_selection_service import select_resorts_from_query
+from src.shared.config import DEFAULT_RESORTS_FILE
 
 
 def _sample_catalog():
@@ -39,7 +40,7 @@ def _sample_catalog():
 
 
 def test_select_resorts_defaults_to_default_scope(monkeypatch):
-    monkeypatch.setattr("src.backend.weather_data_server.load_resort_catalog", lambda path: _sample_catalog())
+    monkeypatch.setattr("src.backend.services.resort_selection_service.load_resort_catalog", lambda path: _sample_catalog())
 
     selected, resorts_file, applied, available, no_match = select_resorts_from_query({})
 
@@ -56,7 +57,7 @@ def test_select_resorts_defaults_to_default_scope(monkeypatch):
 
 
 def test_select_resorts_search_all_ignores_other_filters(monkeypatch):
-    monkeypatch.setattr("src.backend.weather_data_server.load_resort_catalog", lambda path: _sample_catalog())
+    monkeypatch.setattr("src.backend.services.resort_selection_service.load_resort_catalog", lambda path: _sample_catalog())
 
     selected, resorts_file, applied, available, no_match = select_resorts_from_query(
         {
@@ -81,7 +82,7 @@ def test_select_resorts_search_all_ignores_other_filters(monkeypatch):
 
 
 def test_select_resorts_search_filtered_respects_filters(monkeypatch):
-    monkeypatch.setattr("src.backend.weather_data_server.load_resort_catalog", lambda path: _sample_catalog())
+    monkeypatch.setattr("src.backend.services.resort_selection_service.load_resort_catalog", lambda path: _sample_catalog())
 
     selected, resorts_file, applied, _available, no_match = select_resorts_from_query(
         {
@@ -103,7 +104,7 @@ def test_select_resorts_search_filtered_respects_filters(monkeypatch):
 
 
 def test_select_resorts_accepts_multi_value_subregion_and_country(monkeypatch):
-    monkeypatch.setattr("src.backend.weather_data_server.load_resort_catalog", lambda path: _sample_catalog())
+    monkeypatch.setattr("src.backend.services.resort_selection_service.load_resort_catalog", lambda path: _sample_catalog())
 
     selected, resorts_file, applied, _available, no_match = select_resorts_from_query(
         {
@@ -121,7 +122,7 @@ def test_select_resorts_accepts_multi_value_subregion_and_country(monkeypatch):
 
 
 def test_select_resorts_supports_long_form_state_country_and_city(monkeypatch):
-    monkeypatch.setattr("src.backend.weather_data_server.load_resort_catalog", lambda path: _sample_catalog())
+    monkeypatch.setattr("src.backend.services.resort_selection_service.load_resort_catalog", lambda path: _sample_catalog())
 
     selected_state, _file, _applied, _available, no_match_state = select_resorts_from_query(
         {"search": ["utah"], "search_all": ["1"], "include_all": ["1"]}
